requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiogram==3.13.1
python-dotenv==1.0.1
httpx==0.27.2
SQLAlchemy==2.0.36
